+++ port_ocean/core/event_listener/__init__.py
@@ -11,10 +11,17 @@ from port_ocean.core.event_listener.polling import (
     PollingEventListener,
 )
 
+from port_ocean.core.event_listener.once import (
+    OnceEventListenerSettings,
+    OnceEventListener,
+)
+
+
 EventListenerSettingsType = (
     HttpEventListenerSettings
     | KafkaEventListenerSettings
     | PollingEventListenerSettings
+    | OnceEventListenerSettings
 )
 
 __all__ = [
@@ -25,4 +32,6 @@ __all__ = [
     "KafkaEventListenerSettings",
     "PollingEventListener",
     "PollingEventListenerSettings",
+    "OnceEventListener",
+    "OnceEventListenerSettings",
 ]