+++ integrations/gitlab/CHANGELOG.md
@@ -7,6 +7,46 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 <!-- towncrier release notes start -->
 
+0.1.25 (2023-11-03)
+===================
+
+### Improvements
+
+- Bumped ocean version to ^0.4.1 (#1)
+
+
+0.1.24 (2023-11-01)
+===================
+
+### Improvements
+
+- Bumped ocean version to ^0.4.0 and handle ONCE event listener (#1)
+
+
+0.1.23 (2023-10-30)
+===================
+
+### Features
+
+- Added support for project folders (PORT-5060)
+
+
+0.1.22 (2023-10-30)
+===================
+
+### Improvements
+
+- Bumped ocean version to 0.3.2 (#1)
+
+
+0.1.21 (2023-10-29)
+===================
+
+### Features
+
+- Added support for project languages (PORT-4749)
+
+
 0.1.20 (2023-10-24)
 ===================